Renamig and merging classes
