--- app/src/fdroid/java/org/openobservatory/ooniprobe/common/ExceptionManager.java
@@ -1,7 +1,0 @@
-package org.openobservatory.ooniprobe.common;
-
-public class ExceptionManager {
-    public static void logException(Exception e){
-        // does nothing
-    }
-}
