--- app/src/fdroid/java/org/openobservatory/ooniprobe/common/FlavorApplication.java
@@ -1,12 +1,0 @@
-package org.openobservatory.ooniprobe.common;
-
-import android.content.Context;
-
-class FlavorApplication {
-	public static void onCreate(Application app) {
-		// does nothing
-	}
-
-	public static void reloadCrashConsent(Context ctx, PreferenceManager preferenceManager) {
-	}
-}
