+++ app/src/fdroid/java/org/openobservatory/ooniprobe/common/ThirdPartyServices.java
@@ -1,0 +1,29 @@
+package org.openobservatory.ooniprobe.common;
+
+import android.content.Context;
+
+public class ThirdPartyServices {
+    //TODO remove all the non needed methods
+    public static void onCreateApplication(Application app) {
+    }
+
+    public static void register(Application app){
+    }
+
+    public static void reloadConsent(Application app){
+    }
+
+    public static void setToken(String token){
+    }
+
+    public static void initPush(Application app){
+    }
+
+    public static void reloadCrashConsent(Application app) {
+    }
+
+    public static void logException(Exception e){
+        // does nothing
+    }
+
+    }
