+++ app/src/main/java/org/openobservatory/ooniprobe/fragment/onboarding/Onboarding3Fragment.java
@@ -13,7 +13,7 @@ import androidx.fragment.app.Fragment;
 import org.openobservatory.ooniprobe.R;
 import org.openobservatory.ooniprobe.activity.MainActivity;
 import org.openobservatory.ooniprobe.common.Application;
-import org.openobservatory.ooniprobe.common.CountlyManager;
+import org.openobservatory.ooniprobe.common.ThirdPartyServices;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -39,7 +39,7 @@ public class Onboarding3Fragment extends Fragment {
 	@OnClick(R.id.master) void masterClick() {
 		((Application) getActivity().getApplication()).getPreferenceManager().setShowOnboarding(false);
 		((Application) getActivity().getApplication()).getPreferenceManager().acceptDefaultSettings();
-		CountlyManager.reloadConsent((Application) getActivity().getApplication());
+		ThirdPartyServices.reloadConsent((Application) getActivity().getApplication());
 		startActivity(MainActivity.newIntent(getActivity(), R.id.dashboard));
 		getActivity().finish();
 	}
